task: Icon component moved to ui components
